refactor GoogleCloud
